fix: avoid use of Result<_, ()>
The unit type doesn't implement a lot of traits that make it nice to
use as an error, so it's not a great thing to put into a Result's E,
especially not in a `pub` method.

Option<()> is semantically odd, but it has a lot of conveniences, like
`ok_or_else` which is a perfect fit for the "type hole" here.