Fix handling of inverted registered outputs on GAL22V10
As pointed out in https://github.com/simon-frankau/galette/issues/12 ,
the GAL22V10 in registered mode is unusual - the feedback value is
always inverted, even if the output pin value is not. This was not
captured correctly in the code. This is hopefully fixed, along with
some new tests to keep an eye on this annoying corner case.